fix: search param is null when mobile click search btn
