+++ frontend/app/collections/search/page.js
@@ -4,6 +4,7 @@ import { useSearchParams } from 'next/navigation';
 import { useCollections } from '@/context/CollectionsContext';
 import Collections from '@/components/collections/Collections';
 import Card from '@/components/Card';
+import CollectionsSearch from '@/components/CollectionsSearch';
 
 export default function CollectionSearchPage() {
   const { fetchCollections } = useCollections();
@@ -20,6 +21,12 @@ export default function CollectionSearchPage() {
         type="page"
         title={searchQuery === '' ? "Popular collections" : `Search collection: "${searchQuery}"`}
       >
+        <div className="flex justify-end w-full">
+           <div className="w-3/4 max-w-80">
+              <CollectionsSearch />
+           </div>
+        </div>
+       
         <Collections />
       </Card>
