+++ frontend/components/nav/LoggedInComponent.jsx
@@ -39,7 +39,7 @@ const LoggedInComponent = ({ user }) => {
                 </DropdownItem>
                 <hr className="my-2" />
                 <DropdownItem href="/protected/profile" icon={<ProfileIcon />}>Profile</DropdownItem>
-                <DropdownItem href="/protected/collections" icon={<CollectionIcon />}>My Collection</DropdownItem>
+                <DropdownItem href="/protected/collections" icon={<CollectionIcon className="py-2" size={45}/>}>My Collection</DropdownItem>
                 <LogOutDropDownItem />
             </DropdownMenu>
         </>
